ai-email-automation-agent/requirements.txt
streamlit
sentence-transformers
scikit-learn
google-generativeai
